major cleanup, moved information prints to chunk_log, tested push and pull streamers to work with and without fixed delay; second case still has problems with too slow consume period

+++ streamer.c
@@ -90,7 +90,7 @@ extern bool topo_in;
 extern bool topo_keep_best;
 extern bool topo_add_best;
 extern bool autotune_period;
-extern int fixed_playout_delay;
+extern double fixed_playout_delay;
 extern int opmode;
 extern int pullmode;
 
@@ -252,7 +252,7 @@ static void cmdline_parse(int argc, char *argv[])
         else if( strcmp( "topo_keep_best", long_options[option_index].name ) == 0 ) { topo_keep_best = true; }
         else if( strcmp( "topo_add_best", long_options[option_index].name ) == 0 ) { topo_add_best = true; }
         if( strcmp( "autotune_period", long_options[option_index].name ) == 0 ) { autotune_period = (bool) atoi(optarg); }
-        if( strcmp( "fixed_playout_delay", long_options[option_index].name ) == 0 ) { fixed_playout_delay = atoi(optarg); }
+        if( strcmp( "fixed_playout_delay", long_options[option_index].name ) == 0 ) { fixed_playout_delay = atof(optarg); }
         if( strcmp( "pull_earliest", long_options[option_index].name ) == 0 ) { opmode=MODE_PULL;pullmode=PULL_EARLIEST; }
         if( strcmp( "pull_latest", long_options[option_index].name ) == 0 ) { opmode=MODE_PULL;pullmode=PULL_LATEST;
 /*fprintf(stderr,"DEBUG: mode is %d, should be %d\n\n\n",opmode, MODE_PULL); */
